+++ source/idea/idea-data-model/src/ideadatamodel/virtual_desktop/virtual_desktop_model.py
@@ -88,6 +88,7 @@ class DayOfWeek(str, Enum):
 class VirtualDesktopScheduleType(str, Enum):
     WORKING_HOURS = 'WORKING_HOURS'
     STOP_ON_IDLE = 'STOP_ON_IDLE'
+    STOP_ALL_DAY = 'STOP_ALL_DAY'
     START_ALL_DAY = 'START_ALL_DAY'
     CUSTOM_SCHEDULE = 'CUSTOM_SCHEDULE'
     NO_SCHEDULE = 'NO_SCHEDULE'